// src/app/core/interfaces/specification.ts
export interface SpecificationArray {
      productSpecificationId: number;
      productId: number;
      specificationKey: string;
      specificationValue: string ;
      isActive: boolean
}
export interface SpecificationObjArray {
      productSpecificationId: number;
      productId: number;
      specificationKey: string;
      specificationValue: string;
      isActive: boolean
}